% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{Obtain age information by ID number}
\usage{
age(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Age vector obtained by ID numbers.
}
\description{
obtain age information by ID number, to the day.
}
\examples{

id = c(
"652801197305161555", 
"110101840116177"
)
age(id)

}
