% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_by_year.R
\name{age_by_year}
\alias{age_by_year}
\title{Obtain age information by ID number}
\usage{
age_by_year(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Age vector obtained by ID numbers.
}
\description{
obtain age information by ID number, only by year, not the specific date.
}
\examples{

id = c(
"652801197305161555", 
"110101840116177"
)
age_by_year(id)

}
