% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_date.R
\name{birth_date}
\alias{birth_date}
\title{Obtain date of birth information by ID number}
\usage{
birth_date(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Date of birth vector obtained by ID numbers.
}
\description{
obtain date of birth information by ID number.
}
\examples{

id = c(
"652801197305161555", 
"110101841125178"
)
birth_date(id)

}
