% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert18.R
\name{convert18}
\alias{convert18}
\title{Convert the 15 Digits ID Number to 18 Digits}
\usage{
convert18(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
The corresponding 18 digits ID number.
}
\description{
Convert the 15 Digits ID Number to 18 Digits.
}
\examples{
id = c(
  "653127198503161793",
  "652801197305161555",
  "130206202202291545", 
  "110101841125178",
  "12345678",
  "65312a198204181793"
)
convert18(id)

}
