% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leap_year.R
\name{leap_year}
\alias{leap_year}
\title{Calculate whether a given year is a leap year}
\usage{
leap_year(year)
}
\arguments{
\item{year}{A given year, vectorization parameter passing is supported.}
}
\value{
TRUE or FALSE. Return TRUE if it is a leap year, FALSE otherwise.
}
\description{
Calculate whether a given year is a leap year.
}
\examples{
leap_year(2024)

}
