% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdays.R
\name{mdays}
\alias{mdays}
\title{Obtain days of a given year and month.}
\usage{
mdays(month, year = 2022)
}
\arguments{
\item{month}{A given month, such as 2.}

\item{year}{A given year, such as 2022.}
}
\value{
Days of a given year and month.
}
\description{
obtain days of a given year and month.
}
\examples{

mdays(2, 2022)

}
