% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region}
\alias{region}
\title{Obtain region information by ID number}
\usage{
region(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Region vector obtained by ID numbers.
}
\description{
obtain region information by ID number.
}
\examples{

id = c(
"652801197305161555", 
"110101840116177"
)
region(id)

}
