% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ydays.R
\name{ydays}
\alias{ydays}
\title{Calculate the number of days in a specified year}
\usage{
ydays(year = as.integer(format(Sys.Date(), "\%Y")))
}
\arguments{
\item{year}{A given year, the default is current year.}
}
\value{
Days in a specified year.
}
\description{
Calculate the number of days in a specified year.
}
\examples{
ydays(2022)

}
