% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{Get age information from ID number}
\usage{
age(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Age vector obtained from ID numbers.
}
\description{
Get age information from ID number, to the day.
}
\examples{
id = c(
"652801197305161555",
"130206202202291545", 
"110101841125178"
)
age(id)

}
