% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_date.R
\name{birth_date}
\alias{birth_date}
\title{Get date of birth information from ID number}
\usage{
birth_date(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Date of birth vector obtained from ID numbers.
}
\description{
Get date of birth information from ID number.
}
\examples{
id = c(
"653127198503161793",
"652801197305161555",
"130206202202291545", 
"110101841125178",
"12345678",
"65312a198204181793"
)
birth_date(id)

}
