% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_day.R
\name{birth_day}
\alias{birth_day}
\title{Obtain day of birth information by ID number}
\usage{
birth_day(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Day of birth vector obtained by ID numbers.
}
\description{
obtain day of birth information by ID number.
}
\examples{

id = c(
"652801197305161555", 
"110101840116177"
)
birth_day(id)

}
