% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_day.R
\name{birth_day}
\alias{birth_day}
\title{Get day of birth information from ID number}
\usage{
birth_day(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Day of birth vector obtained from ID numbers.
}
\description{
Get day of birth information from ID number.
}
\examples{
id = c(
"652801197305161555",
"130206202202291545", 
"110101841125178"
)
birth_day(id)

}
