% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_month.R
\name{birth_month}
\alias{birth_month}
\title{Obtain month of birth information by ID number}
\usage{
birth_month(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Month of birth vector obtained by ID numbers.
}
\description{
obtain month of birth information by ID number.
}
\examples{

id = c(
"652801197305161555", 
"110101840116177"
)
birth_month(id)

}
