% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_month.R
\name{birth_month}
\alias{birth_month}
\title{Get month of birth information from ID number}
\usage{
birth_month(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Month of birth vector obtained from ID numbers.
}
\description{
Get month of birth information from ID number.
}
\examples{
id = c(
"652801197305161555",
"130206202202291545", 
"110101841125178"
)
birth_month(id)

}
