% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_year.R
\name{birth_year}
\alias{birth_year}
\title{Get year of birth information from ID number}
\usage{
birth_year(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Year of birth vector obtained from ID numbers.
}
\description{
Get year of birth information from ID number.
}
\examples{
id = c(
"652801197305161555",
"130206202202291545", 
"110101841125178"
)
birth_year(id)

}
