% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_id.R
\name{check_id}
\alias{check_id}
\title{Check whether the ID number is valid}
\usage{
check_id(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
A vector of TRUE or FALSE.
}
\description{
Check whether the ID number is valid.
}
\examples{
id = c(
  "653127198503161793",
  "652801197305161555",
  "130206202202291545", 
  "110101841125178",
  "12345678",
  "65312a198204181793"
)
check_id(id)

}
