% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_id.R
\name{check_id}
\alias{check_id}
\title{Check the ID number for logical errors}
\usage{
check_id(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
A vector of TRUE or FALSE.
}
\description{
Check the ID number for logical errors.
}
\examples{
id = c(
"652801197305161555",
"130206202202291545", 
"110101841125178"
)
check_id(id)

}
