% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnid_info.R
\name{cnid_info}
\alias{cnid_info}
\title{Obtain full information by ID number}
\usage{
cnid_info(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
A list about date of birth, age, gender, etc. 
obtained by ID number.
}
\description{
obtain full information by ID number.
}
\examples{

id = c(
"652801197305161555", 
"110101841125178"
)
cnid_info(id)

}
