% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnid_info.R
\name{cnid_info}
\alias{cnid_info}
\title{Get full information from ID number}
\usage{
cnid_info(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
A list about date of birth, age, gender, etc.
obtained from ID number.
}
\description{
Get full information from ID number.
}
\examples{
id = c(
"653127198503161793",
"652801197305161555",
"130206202202291545", 
"110101841125178",
"12345678",
"65312a198204181793"
)
cnid_info(id)

}
