% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{constellation}
\alias{constellation}
\title{Obtain constellation information by ID number}
\usage{
constellation(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Constellation vector obtained by ID numbers.
}
\description{
obtain constellation information by ID number.
}
\examples{

id = c(
"652801197305161555", 
"110101841125178"
)
constellation(id)

}
