% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender.R
\name{gender}
\alias{gender}
\title{Obtain gender information by ID number}
\usage{
gender(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Gender vector obtained by ID numbers.
}
\description{
obtain gender information by ID number.
}
\examples{

id = c(
"652801197305161555", 
"110101840116177"
)
gender(id)

}
