% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender.R
\name{gender}
\alias{gender}
\title{Get gender information from ID number}
\usage{
gender(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Gender vector obtained from ID numbers.
}
\description{
Get gender information from ID number.
}
\examples{
id = c(
  "653127198503161793",
  "652801197305161555",
  "130206202202291545", 
  "110101841125178",
  "12345678",
  "65312a198204181793"
)
gender(id)

}
