% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdays.R
\name{mdays}
\alias{mdays}
\title{Calculate the number of days in a given year and month}
\usage{
mdays(year, month)
}
\arguments{
\item{year}{A given year.}

\item{month}{A given month, vectorization parameter passing is supported.}
}
\value{
Days in a given year and month.
}
\description{
Calculate the number of days in a given year and month.
}
\examples{
mdays(2025, 1:12)

}
