% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdays.R
\name{mdays}
\alias{mdays}
\title{Calculate the number of days in a specified year and month}
\usage{
mdays(month, year = as.integer(format(Sys.Date(), "\%Y")))
}
\arguments{
\item{month}{A given month, such as 2.}

\item{year}{A given year, the default is current year.}
}
\value{
Days in a specified year and month.
}
\description{
Calculate the number of days in a specified year and month.
}
\examples{
mdays(2, 2022)

}
