% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region}
\alias{region}
\title{Get region information from ID number}
\usage{
region(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Region vector obtained from ID numbers.
}
\description{
Get region information from ID number.
}
\examples{
id = c(
  "653127198503161793",
  "652801197305161555",
  "130206202202291545", 
  "110101841125178",
  "12345678",
  "65312a198204181793"
)
region(id)

}
