% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ydays.R
\name{ydays}
\alias{ydays}
\title{Calculate the number of days in a given year}
\usage{
ydays(year)
}
\arguments{
\item{year}{A given year, vectorization parameter passing is supported.}
}
\value{
Days in a given year.
}
\description{
Calculate the number of days in a given year.
}
\examples{
ydays(2025)

}
