% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zodiac.R
\name{zodiac}
\alias{zodiac}
\title{Obtain zodiac information by ID number}
\usage{
zodiac(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Zodiac vector obtained by ID numbers.
}
\description{
obtain zodiac information by ID number.
}
\examples{

id = c(
"652801197305161555", 
"110101840116177"
)
zodiac(id)

}
