% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zodiac.R
\name{zodiac}
\alias{zodiac}
\title{Get zodiac information from ID number}
\usage{
zodiac(id)
}
\arguments{
\item{id}{A vector of ID numbers.}
}
\value{
Zodiac vector obtained from ID numbers.
}
\description{
Get zodiac information from ID number.
}
\examples{
id = c(
"652801197305161555",
"130206202202291545", 
"110101841125178"
)
zodiac(id)

}
